delete card reducer
Удалил card-ветку в redux, изменил ветку column, объединил запросы по колонкам и карточкам в один запрос.
