+++ src/redux/store.js
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import authSlice from './auth/authSlice'
 import boardSlice from './board/boardSlice'
-import cardSlice from './card/cardSlice'
 import columnSlice from './column/columnSlice'
 
 
@@ -9,7 +8,6 @@ export const store = configureStore({
     reducer: {
         auth: authSlice,
         board: boardSlice,
-        column: columnSlice,
-        card: cardSlice
+        column: columnSlice
     }
 })
